requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
psycopg2==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
